Type time buffers as uint32

+++ firmware/dwm1001_dev_source_code/development/dedicated_time_protocol/main.c
@@ -131,7 +131,7 @@ int txCounter = 0; // debugging purpose
  * Each elem i is the timestamp where node NODE_ID rx the transmission 
  * from node i.
  */
-uint8 timeOwn[NUM_STAMPS_PER_NODE*N];
+uint32 timeOwn[NUM_STAMPS_PER_NODE*N];
 
 /**
  * Times that other nodes (not NODE_ID) stamped.
@@ -139,7 +139,7 @@ uint8 timeOwn[NUM_STAMPS_PER_NODE*N];
  * Each elem i is the timestamp where node i rx the transmission from 
  * node NODE_ID.
  */
-uint8 timeOthers[NUM_STAMPS_PER_NODE*N];
+uint32 timeOthers[NUM_STAMPS_PER_NODE*N];
 
 /** Default header */
 uint8 header[10] = {0x41, 0x88, 0, 0xCA, 0xDE, 'W', 'A', 'V', 'E', 0xE0};
